+++ CHANGELOG.md
@@ -31,8 +31,8 @@ v1.10 (not yet released)
   * [`select-cursor-char`]
   * [`utf8-bom`]
 * Added an optional *exitcode* argument to the [`quit`] command.
-* Added `color`, `env`, `option` and `wsplit` arguments to the [`show`]
-  command.
+* Added `color`, `command`, `env`, `option`, `search` and `wsplit`
+  arguments to the [`show`] command.
 * Added support for the `\e` escape sequence in [double-quoted] command
   arguments.
 * Added syntax highlighting for Lisp and Scheme files.